system test issue #36399: It display the whole shift as overtime when I add overtime at the begining of the shift

+++ Teleopti.Ccc.Web/Teleopti.Ccc.Web/Areas/Anywhere/Core/GroupScheduleShiftViewModel.cs
@@ -7,8 +7,8 @@ namespace Teleopti.Ccc.Web.Areas.Anywhere.Core
 		public string PersonId { get; set; }
 		public string Name { get; set; }
 		public string Date { get; set; }
-		public int WorkTimeMinutes { get; set; }
-		public int ContractTimeMinutes { get; set; }
+		public double WorkTimeMinutes { get; set; }
+		public double ContractTimeMinutes { get; set; }
 
 		public IEnumerable<GroupScheduleLayerViewModel> Projection { get; set; }
 		public bool IsFullDayAbsence { get; set; }
